Using the Context API for state Management
The fetch data is making in Server Components provide to fast render. And if the URL is changed and the server component has to be  rendered. The Context API just works for client component because hooks can be used in the Client Component. Client and Server components communication just provides wit ContextAPI